# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "APS Token Management API",
        "title": "APS Viewer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/aps/buckets": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get list of all buckets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Bucket"
                ],
                "summary": "Get buckets list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.APSBucket"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "新しいAPSバケットを作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Bucket"
                ],
                "summary": "APSバケット作成",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.APSBucket"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/buckets/{bucketKey}": {
            "delete": {
                "description": "指定されたバケットを削除します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Bucket"
                ],
                "summary": "APSバケット削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "バケットキー",
                        "name": "bucketKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/buckets/{bucketKey}/details": {
            "get": {
                "description": "指定されたバケットの詳細情報を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Bucket"
                ],
                "summary": "APSバケット詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "バケットキー",
                        "name": "bucketKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.APSBucketDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_bucket.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/buckets/{bucketKey}/objects/signeds3upload": {
            "post": {
                "description": "オブジェクトをS3へ保存するためのS3署名付きURLを取得します",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "S3署名付きURLの取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "バケットキー",
                        "name": "bucketKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "アップロードするファイル",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "パート数",
                        "name": "parts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.APSObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/buckets/{bucketKey}/objects/upload": {
            "post": {
                "description": "S3署名付きURLを取得してオブジェクトをアップロードするシーケンスを実行します",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "APSオブジェクトのアップロードシーケンス",
                "parameters": [
                    {
                        "type": "string",
                        "description": "バケットキー",
                        "name": "bucketKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "アップロードするファイル",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "パート数",
                        "name": "parts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/buckets/{bucketKey}/objects/{objectKey}/signeds3upload": {
            "post": {
                "description": "S3へアップロードしたオブジェクトの作成を完了します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "APSオブジェクトの作成完了",
                "parameters": [
                    {
                        "type": "string",
                        "description": "バケットキー",
                        "name": "bucketKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "オブジェクトキー",
                        "name": "objectKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "アップロードキー",
                        "name": "uploadKey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.APSObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/objects/signeds3upload": {
            "put": {
                "description": "S3署名付きURLを使用してオブジェクトをS3へアップロードします",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "S3署名付きURLを使用したオブジェクトのアップロード",
                "parameters": [
                    {
                        "type": "string",
                        "description": "署名付きURL",
                        "name": "signedURL",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "アップロードするファイルのバイナリデータ",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/objects/{objectId}/base64urn": {
            "get": {
                "description": "オブジェクトIDをBase64エンコードしたURNを生成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "オブジェクトURNのBase64エンコード",
                "parameters": [
                    {
                        "type": "string",
                        "description": "オブジェクトID",
                        "name": "objectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/objects/{objectId}/translate": {
            "post": {
                "description": "オブジェクトの翻訳ジョブを作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "APSオブジェクトの翻訳ジョブ作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "オブジェクトID",
                        "name": "objectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "オブジェクトキー",
                        "name": "objectKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TranslateJobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/objects/{urn}/status": {
            "get": {
                "description": "翻訳ジョブの進捗状況を確認します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Object"
                ],
                "summary": "翻訳ジョブのステータス確認",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64エンコードされたURN",
                        "name": "urn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TranslationStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/aps_object.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/aps/token": {
            "post": {
                "description": "2-legged認証でAPSアクセストークンを取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APS Token"
                ],
                "summary": "APSトークン取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.APSToken"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "aps_bucket.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "aps_object.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "domain.APSBucket": {
            "type": "object",
            "properties": {
                "bucketKey": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "integer"
                },
                "policyKey": {
                    "type": "string"
                }
            }
        },
        "domain.APSBucketDetail": {
            "type": "object",
            "properties": {
                "bucketKey": {
                    "type": "string"
                },
                "bucketOwner": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Permission"
                    }
                },
                "policyKey": {
                    "type": "string"
                }
            }
        },
        "domain.APSObject": {
            "type": "object",
            "properties": {
                "bucketKey": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "objectId": {
                    "type": "string"
                },
                "objectKey": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "uploadExpiration": {
                    "type": "string"
                },
                "uploadKey": {
                    "type": "string"
                },
                "urlExpiration": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "domain.APSToken": {
            "description": "APSトークンレスポンス",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3599
                },
                "refresh_token": {
                    "type": "string",
                    "example": ""
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "domain.Children": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Resource"
                    }
                },
                "guid": {
                    "type": "string"
                },
                "hasThumbnail": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Message"
                    }
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "domain.Derivative": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Children"
                    }
                },
                "hasThumbnail": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Message"
                    }
                },
                "name": {
                    "type": "string"
                },
                "outputType": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.Message": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "domain.Permission": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string"
                },
                "authId": {
                    "type": "string"
                }
            }
        },
        "domain.Resource": {
            "type": "object",
            "properties": {
                "guid": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "resolution": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "urn": {
                    "type": "string"
                }
            }
        },
        "domain.TranslateJobResponse": {
            "type": "object",
            "properties": {
                "acceptedJobs": {
                    "type": "object",
                    "properties": {
                        "output": {
                            "type": "object",
                            "properties": {
                                "formats": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "type": {
                                                "type": "string"
                                            },
                                            "views": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "result": {
                    "type": "string"
                },
                "urn": {
                    "type": "string"
                }
            }
        },
        "domain.TranslationStatus": {
            "type": "object",
            "properties": {
                "derivatives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Derivative"
                    }
                },
                "hasThumbnail": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "urn": {
                    "type": "string"
                }
            }
        }
    }
}